removed comments - otherwise app wouldn't load when navigating to localhost:3000/login/rootview.
Before removal, a 
typeerror-object-doesnt-support-this-property-or-method was thrown when 
loading the app, refering to a line in application.html.erb (in 
views/layouts folder). It was referring to the stylesheet line, however 
fixing the js file worked. Thanks to a SO answer suggesting to remove 
the comments from the js file - indeed turbolinks was required, so it 
makes some sense.