Generated new 'posts' controller.
Was getting errors that "PostController" was not defined. I tried to do 
explicit redirects (:controller => <mycontrollername>...) but I could 
not get it working so far. I followed the implications given by the 
errors, and created the things the browser was looking for. Ideally 
though, I would like to call controllers whatever I want, so will need 
to look into redirects more.