Happy accident! Kept a nice stylesheet from generating a (--skip) blogs scaffold, after deleting all other generated files.
I realised I did not want to generate a scaffold, or not yet, but I
observed that the page style changed (for the better!) and thought it
must be stylesheet. I removed all other files resulting from generating
the scaffold, but for the scss one, and get to keep the nicer look.